Move test with GrammarKit to GrammarKit tests
