+++ sqldelight-gradle-plugin/src/dockerTest/kotlin/app/cash/sqldelight/dialect/DialectIntegrationTests.kt
@@ -51,15 +51,6 @@ class DialectIntegrationTests {
     val result = runner.build()
     Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
   }
-
-  @Test fun `dialect accepts version catalog dependency`() {
-    val runner = GradleRunner.create()
-      .withCommonConfiguration(File("src/test/integration-catalog"))
-      .withArguments("clean", "check", "--stacktrace")
-
-    val result = runner.build()
-    Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
-  }
 }
 
 private fun GradleRunner.withCommonConfiguration(projectRoot: File): GradleRunner {
