+++ sqldelight-gradle-plugin/src/test/kotlin/app/cash/sqldelight/dialect/DialectIntegrationTests.kt
@@ -1,12 +1,10 @@
 package app.cash.sqldelight.dialect
 
-import app.cash.sqldelight.UsesGrammarKit
 import app.cash.sqldelight.assertions.FileSubject
 import app.cash.sqldelight.withCommonConfiguration
 import com.google.common.truth.Truth
 import org.gradle.testkit.runner.GradleRunner
 import org.junit.Test
-import org.junit.experimental.categories.Category
 import java.io.File
 
 class DialectIntegrationTests {
@@ -47,17 +45,6 @@ class DialectIntegrationTests {
     val runner = GradleRunner.create()
       .withCommonConfiguration(File("src/test/integration-multi-dialect-modules"))
       .withArguments("clean", "build", "--stacktrace")
-
-    val result = runner.build()
-    Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
-  }
-
-  @Category(UsesGrammarKit::class)
-  @Test
-  fun customFunctionDialect() {
-    val runner = GradleRunner.create()
-      .withCommonConfiguration(File("src/test/custom-dialect"))
-      .withArguments("clean", "compileTestKotlin", "--stacktrace")
 
     val result = runner.build()
     Truth.assertThat(result.output).contains("BUILD SUCCESSFUL")
